# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export type AuthUserRole = typeof AuthUserRole[keyof typeof AuthUserRole];


export const AuthUserRole = {
  farmer: 'farmer',
  fpo: 'fpo',
  admin: 'admin',
} as const;

export interface AuthUser {
  id: string;
  role: AuthUserRole;
  displayName: string;
  identity: string;
}

export interface AuthSession {
  authenticated: boolean;
  user: AuthUser | null;
}

export type AuthSignInInputRole = typeof AuthSignInInputRole[keyof typeof AuthSignInInputRole];


export const AuthSignInInputRole = {
  farmer: 'farmer',
  fpo: 'fpo',
  admin: 'admin',
} as const;

export interface AuthSignInInput {
  role: AuthSignInInputRole;
  /** @minLength 1 */
  identity: string;
  /** @minLength 1 */
  password: string;
  rememberMe?: boolean;
}

export interface HealthStatus {
  status: string;
}

export type FpoStatus = typeof FpoStatus[keyof typeof FpoStatus];


export const FpoStatus = {
  Pending: 'Pending',
  Under_Review: 'Under Review',
  Verified: 'Verified',
  Rejected: 'Rejected',
} as const;

export type DocumentStatusStatus = typeof DocumentStatusStatus[keyof typeof DocumentStatusStatus];


export const DocumentStatusStatus = {
  Pending: 'Pending',
  Under_Review: 'Under Review',
  Verified: 'Verified',
  Rejected: 'Rejected',
} as const;

export interface DocumentStatus {
  id: string;
  name: string;
  status: DocumentStatusStatus;
  /** @nullable */
  reason?: string | null;
  /** @nullable */
  fileName?: string | null;
}

export interface Fpo {
  id: string;
  name: string;
  registrationNumber: string;
  state: string;
  district: string;
  block: string;
  villages: string[];
  crops: string[];
  memberCount: number;
  totalArea: number;
  status: FpoStatus;
  description: string;
  contactName: string;
  contactMobile: string;
  documents: DocumentStatus[];
}

export interface FpoOnboardingInput {
  name: string;
  state: string;
  district: string;
  block: string;
  contactName: string;
  mobile: string;
  email: string;
  registrationNumber: string;
  incorporationDate: string;
  pan: string;
  /** @nullable */
  gst?: string | null;
  legalAct: string;
  crops?: string[];
  documents?: DocumentStatus[];
}

export type FpoOnboarding = Fpo & ({
  step: number;
  submittedAt: string;
  legalAct?: string;
  incorporationDate?: string;
  pan?: string;
  /** @nullable */
  gst?: string | null;
  email?: string;
});

export interface Member {
  id: string;
  farmerId: string;
  name: string;
  village: string;
  mobile: string;
  landholding: number;
  crops: string[];
}

export interface MemberInput {
  name: string;
  village: string;
  mobile: string;
  landholding: number;
  crops: string[];
}

export interface JoinRequestInput {
  fpoId: string;
  name: string;
  village: string;
  mobile: string;
  landholding: number;
  crops: string[];
}

export type JoinRequestStatus = typeof JoinRequestStatus[keyof typeof JoinRequestStatus];


export const JoinRequestStatus = {
  Pending: 'Pending',
  Accepted: 'Accepted',
  Rejected: 'Rejected',
} as const;

export type JoinRequest = Member & {
  requestId: string;
  fpoId: string;
  status: JoinRequestStatus;
  requestedAt: string;
};

export type FpoReviewAuditLogItem = {
  id: string;
  actor: string;
  action: string;
  at: string;
  /** @nullable */
  detail?: string | null;
};

export interface FpoReview {
  fpo: Fpo;
  auditLog: FpoReviewAuditLogItem[];
}

export type FpoDecisionInputDecision = typeof FpoDecisionInputDecision[keyof typeof FpoDecisionInputDecision];


export const FpoDecisionInputDecision = {
  approve: 'approve',
  reject: 'reject',
} as const;

export interface FpoDecisionInput {
  decision: FpoDecisionInputDecision;
  /** @nullable */
  reason?: string | null;
}

export type ListFposParams = {
search?: string;
state?: string;
district?: string;
crop?: string;
status?: string;
};

export type ListAdminFposParams = {
status?: string;
state?: string;
crop?: string;
};