# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: fpos
    description: FPO discovery and management
  - name: onboarding
    description: FPO onboarding workflow
  - name: members
    description: FPO member directory
  - name: join-requests
    description: Farmer membership requests
  - name: admin
    description: FPO review operations
  - name: auth
    description: Role-based authentication and sessions
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/sign-in:
    post:
      operationId: signIn
      tags: [auth]
      summary: Sign in to a role workspace
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthSignInInput" }
      responses:
        "200":
          description: Authenticated session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "400": { description: Invalid sign-in input }
        "401": { description: Invalid credentials }
  /auth/session:
    get:
      operationId: getAuthSession
      tags: [auth]
      summary: Get the current session
      responses:
        "200":
          description: Current session state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
  /auth/sign-out:
    post:
      operationId: signOut
      tags: [auth]
      summary: Sign out the current session
      responses:
        "204": { description: Signed out }
  /fpos:
    get:
      operationId: listFpos
      tags: [fpos]
      summary: List FPOs
      security:
        - SessionCookie: []
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
        - name: district
          in: query
          schema: { type: string }
        - name: crop
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: FPO list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Fpo" }
  /fpos/{fpoId}:
    get:
      operationId: getFpo
      tags: [fpos]
      summary: Get an FPO profile
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/FpoId"
      responses:
        "200":
          description: FPO profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Fpo" }
        "404": { description: FPO not found }
  /fpos/onboarding:
    post:
      operationId: submitFpoOnboarding
      tags: [onboarding]
      summary: Submit FPO onboarding details
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FpoOnboardingInput" }
      responses:
        "201":
          description: Onboarding submission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FpoOnboarding" }
  /fpos/{fpoId}/onboarding:
    get:
      operationId: getFpoOnboarding
      tags: [onboarding]
      summary: Get onboarding status
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/FpoId"
      responses:
        "200":
          description: Onboarding status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FpoOnboarding" }
  /fpos/{fpoId}/members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List FPO members
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/FpoId"
      responses:
        "200":
          description: Member list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Member" }
    post:
      operationId: addMember
      tags: [members]
      summary: Add a farmer to an FPO
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/FpoId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MemberInput" }
      responses:
        "201":
          description: Created member
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Member" }
  /join-requests:
    post:
      operationId: createJoinRequest
      tags: [join-requests]
      summary: Request to join an FPO
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JoinRequestInput" }
      responses:
        "201":
          description: Join request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JoinRequest" }
  /join-requests/{requestId}:
    get:
      operationId: getJoinRequest
      tags: [join-requests]
      summary: Get join request status
      security:
        - SessionCookie: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Join request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JoinRequest" }
  /admin/fpos:
    get:
      operationId: listAdminFpos
      tags: [admin]
      summary: List FPOs for verification
      security:
        - SessionCookie: []
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
        - name: crop
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Admin FPO list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Fpo" }
  /admin/fpos/{fpoId}:
    get:
      operationId: getAdminFpo
      tags: [admin]
      summary: Get FPO review details
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/FpoId"
      responses:
        "200":
          description: FPO review details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FpoReview" }
  /admin/fpos/{fpoId}/decision:
    post:
      operationId: decideFpo
      tags: [admin]
      summary: Approve or reject an FPO
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/FpoId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FpoDecisionInput" }
      responses:
        "200":
          description: Updated FPO
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Fpo" }
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: agri_session
  parameters:
    FpoId:
      name: fpoId
      in: path
      required: true
      schema: { type: string }
  schemas:
    AuthUser:
      type: object
      required: [id, role, displayName, identity]
      properties:
        id: { type: string }
        role: { type: string, enum: [farmer, fpo, admin] }
        displayName: { type: string }
        identity: { type: string }
    AuthSession:
      type: object
      required: [authenticated, user]
      properties:
        authenticated: { type: boolean }
        user: { oneOf: [{ $ref: "#/components/schemas/AuthUser" }, { type: "null" }] }
    AuthSignInInput:
      type: object
      required: [role, identity, password]
      properties:
        role: { type: string, enum: [farmer, fpo, admin] }
        identity: { type: string, minLength: 1 }
        password: { type: string, minLength: 1 }
        rememberMe: { type: boolean, default: false }
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Fpo:
      type: object
      required: [id, name, registrationNumber, state, district, block, villages, crops, memberCount, totalArea, status, description, contactName, contactMobile, documents]
      properties:
        id: { type: string }
        name: { type: string }
        registrationNumber: { type: string }
        state: { type: string }
        district: { type: string }
        block: { type: string }
        villages:
          type: array
          items: { type: string }
        crops:
          type: array
          items: { type: string }
        memberCount: { type: number }
        totalArea: { type: number }
        status: { type: string, enum: [Pending, Under Review, Verified, Rejected] }
        description: { type: string }
        contactName: { type: string }
        contactMobile: { type: string }
        documents:
          type: array
          items: { $ref: "#/components/schemas/DocumentStatus" }
    DocumentStatus:
      type: object
      required: [id, name, status]
      properties:
        id: { type: string }
        name: { type: string }
        status: { type: string, enum: [Pending, Under Review, Verified, Rejected] }
        reason: { type: ["string", "null"] }
        fileName: { type: ["string", "null"] }
    FpoOnboardingInput:
      type: object
      required: [name, state, district, block, contactName, mobile, email, registrationNumber, incorporationDate, pan, legalAct]
      properties:
        name: { type: string }
        state: { type: string }
        district: { type: string }
        block: { type: string }
        contactName: { type: string }
        mobile: { type: string }
        email: { type: string }
        registrationNumber: { type: string }
        incorporationDate: { type: string }
        pan: { type: string }
        gst: { type: ["string", "null"] }
        legalAct: { type: string }
        crops:
          type: array
          items: { type: string }
        documents:
          type: array
          items: { $ref: "#/components/schemas/DocumentStatus" }
    FpoOnboarding:
      allOf:
        - $ref: "#/components/schemas/Fpo"
        - type: object
          required: [step, submittedAt]
          properties:
            step: { type: number }
            submittedAt: { type: string }
            legalAct: { type: string }
            incorporationDate: { type: string }
            pan: { type: string }
            gst: { type: ["string", "null"] }
            email: { type: string }
    Member:
      type: object
      required: [id, farmerId, name, village, mobile, landholding, crops]
      properties:
        id: { type: string }
        farmerId: { type: string }
        name: { type: string }
        village: { type: string }
        mobile: { type: string }
        landholding: { type: number }
        crops:
          type: array
          items: { type: string }
    MemberInput:
      type: object
      required: [name, village, mobile, landholding, crops]
      properties:
        name: { type: string }
        village: { type: string }
        mobile: { type: string }
        landholding: { type: number }
        crops:
          type: array
          items: { type: string }
    JoinRequestInput:
      type: object
      required: [fpoId, name, village, mobile, landholding, crops]
      properties:
        fpoId: { type: string }
        name: { type: string }
        village: { type: string }
        mobile: { type: string }
        landholding: { type: number }
        crops:
          type: array
          items: { type: string }
    JoinRequest:
      allOf:
        - $ref: "#/components/schemas/Member"
        - type: object
          required: [requestId, fpoId, status, requestedAt]
          properties:
            requestId: { type: string }
            fpoId: { type: string }
            status: { type: string, enum: [Pending, Accepted, Rejected] }
            requestedAt: { type: string }
    FpoReview:
      type: object
      required: [fpo, auditLog]
      properties:
        fpo: { $ref: "#/components/schemas/Fpo" }
        auditLog:
          type: array
          items:
            type: object
            required: [id, actor, action, at]
            properties:
              id: { type: string }
              actor: { type: string }
              action: { type: string }
              at: { type: string }
              detail: { type: ["string", "null"] }
    FpoDecisionInput:
      type: object
      required: [decision]
      properties:
        decision: { type: string, enum: [approve, reject] }
        reason: { type: ["string", "null"] }