# lib/api-zod/src/generated/types/authSignInInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthSignInInputRole } from './authSignInInputRole';

export interface AuthSignInInput {
  role: AuthSignInInputRole;
  /** @minLength 1 */
  identity: string;
  /** @minLength 1 */
  password: string;
  rememberMe?: boolean;
}